requirements.txt
Django==1.8.4
Markdown==2.6.2
django-filter==0.11.0
djangorestframework==3.2.4
wsgiref==0.1.2
